Extract descriptor related functionality
