+++ hotham/src/resources/render_context.rs
@@ -19,6 +19,7 @@ use crate::{
     rendering::{
         buffer::Buffer,
         camera::Camera,
+        descriptors::Descriptors,
         frame::Frame,
         image::Image,
         scene_data::{SceneData, SceneParams},
@@ -63,7 +64,7 @@ pub struct RenderContext {
     pub views: Vec<xr::View>,
     pub last_frame_time: Instant,
     pub frame_index: usize,
-    pub descriptors: Descriptors,
+    pub(crate) descriptors: Descriptors,
 }
 
 impl RenderContext {
@@ -861,136 +862,36 @@ fn create_pipeline_layout(
     .map_err(|e| e.into())
 }
 
-/// A wrapper around all the various bits of descriptor functionality
-#[derive(Clone, Debug)]
-pub struct Descriptors {
-    pub layout: vk::DescriptorSetLayout,
-    pub set: vk::DescriptorSet,
-    pub pool: vk::DescriptorPool,
-}
+#[cfg(test)]
+mod tests {
+    use super::*;
 
-impl Descriptors {
-    pub unsafe fn new(vulkan_context: &VulkanContext) -> Self {
-        // First, create a pool.
-        let pool = create_descriptor_pool(&vulkan_context.device);
+    #[test]
+    pub fn render_context_smoke_test() {
+        let vulkan_context = VulkanContext::testing().unwrap();
+        let resolution = vk::Extent2D {
+            height: 800,
+            width: 800,
+        };
+        // Create an image with vulkan_context
+        let image = vulkan_context
+            .create_image(
+                COLOR_FORMAT,
+                &resolution,
+                vk::ImageUsageFlags::COLOR_ATTACHMENT | vk::ImageUsageFlags::TRANSFER_SRC,
+                2,
+                1,
+            )
+            .unwrap();
+        vulkan_context
+            .set_debug_name(vk::ObjectType::IMAGE, image.handle.as_raw(), "Screenshot")
+            .unwrap();
 
-        // Then create a layout.
-        let layout = create_descriptor_layouts(&vulkan_context.device);
+        let swapchain = Swapchain {
+            images: vec![image.handle],
+            resolution,
+        };
 
-        // Finally, allocate the shared descriptor set.
-        let set = allocate_descriptor_set(vulkan_context, pool, layout);
-
-        Self { layout, set, pool }
+        RenderContext::new_from_swapchain(&vulkan_context, &swapchain).unwrap();
     }
-}
-
-unsafe fn allocate_descriptor_set(
-    vulkan_context: &VulkanContext,
-    pool: vk::DescriptorPool,
-    layout: vk::DescriptorSetLayout,
-) -> vk::DescriptorSet {
-    let mut descriptor_counts =
-        vk::DescriptorSetVariableDescriptorCountAllocateInfo::builder().descriptor_counts(&[1000]);
-    let set = vulkan_context
-        .device
-        .allocate_descriptor_sets(
-            &vk::DescriptorSetAllocateInfo::builder()
-                .descriptor_pool(pool)
-                .set_layouts(std::slice::from_ref(&layout))
-                .push_next(&mut descriptor_counts),
-        )
-        .unwrap()[0];
-    set
-}
-
-unsafe fn create_descriptor_layouts(device: &ash::Device) -> vk::DescriptorSetLayout {
-    let bindings = [
-        // Draw Data
-        vk::DescriptorSetLayoutBinding {
-            binding: 0,
-            descriptor_type: vk::DescriptorType::STORAGE_BUFFER,
-            stage_flags: vk::ShaderStageFlags::VERTEX | vk::ShaderStageFlags::COMPUTE,
-            descriptor_count: 1,
-            ..Default::default()
-        },
-        // Models
-        vk::DescriptorSetLayoutBinding {
-            binding: 1,
-            descriptor_type: vk::DescriptorType::STORAGE_BUFFER,
-            stage_flags: vk::ShaderStageFlags::VERTEX | vk::ShaderStageFlags::COMPUTE,
-            descriptor_count: 1,
-            ..Default::default()
-        },
-        // Materials
-        vk::DescriptorSetLayoutBinding {
-            binding: 2,
-            descriptor_type: vk::DescriptorType::STORAGE_BUFFER,
-            stage_flags: vk::ShaderStageFlags::FRAGMENT,
-            descriptor_count: 1,
-            ..Default::default()
-        },
-        // Draw Calls
-        vk::DescriptorSetLayoutBinding {
-            binding: 3,
-            descriptor_type: vk::DescriptorType::STORAGE_BUFFER,
-            stage_flags: vk::ShaderStageFlags::COMPUTE,
-            descriptor_count: 1,
-            ..Default::default()
-        },
-        // Textures
-        vk::DescriptorSetLayoutBinding {
-            binding: 4,
-            descriptor_type: vk::DescriptorType::COMBINED_IMAGE_SAMPLER,
-            stage_flags: vk::ShaderStageFlags::FRAGMENT,
-            descriptor_count: 1000,
-            ..Default::default()
-        },
-    ];
-
-    let flags = vk::DescriptorBindingFlags::PARTIALLY_BOUND
-        | vk::DescriptorBindingFlags::VARIABLE_DESCRIPTOR_COUNT
-        | vk::DescriptorBindingFlags::UPDATE_AFTER_BIND;
-    let descriptor_flags = [
-        vk::DescriptorBindingFlags::empty(),
-        vk::DescriptorBindingFlags::empty(),
-        vk::DescriptorBindingFlags::empty(),
-        vk::DescriptorBindingFlags::empty(),
-        flags,
-    ];
-    let mut binding_flags = vk::DescriptorSetLayoutBindingFlagsCreateInfoEXT::builder()
-        .binding_flags(&descriptor_flags);
-
-    let shared_layout = device
-        .create_descriptor_set_layout(
-            &vk::DescriptorSetLayoutCreateInfo::builder()
-                .bindings(&bindings)
-                .push_next(&mut binding_flags)
-                .flags(vk::DescriptorSetLayoutCreateFlags::UPDATE_AFTER_BIND_POOL),
-            None,
-        )
-        .unwrap();
-
-    shared_layout
-}
-
-unsafe fn create_descriptor_pool(device: &ash::Device) -> vk::DescriptorPool {
-    let pool_sizes = [
-        vk::DescriptorPoolSize {
-            ty: vk::DescriptorType::STORAGE_BUFFER,
-            descriptor_count: 100,
-        },
-        vk::DescriptorPoolSize {
-            ty: vk::DescriptorType::COMBINED_IMAGE_SAMPLER,
-            descriptor_count: 10_000,
-        },
-    ];
-    device
-        .create_descriptor_pool(
-            &vk::DescriptorPoolCreateInfo::builder()
-                .pool_sizes(&pool_sizes)
-                .max_sets(1000)
-                .flags(vk::DescriptorPoolCreateFlags::UPDATE_AFTER_BIND),
-            None,
-        )
-        .unwrap()
 }
